\name{bwe-package}
\alias{bwe-package}
\alias{bwe}
\docType{package}
\title{
\packageTitle{bwe}
}
\description{
\packageDescription{bwe}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bwe}
\packageIndices{bwe}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bwe}

Maintainer: \packageMaintainer{bwe}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
