% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_anchor_word_embeddings.R
\name{double_anchor_word_embeddings}
\alias{double_anchor_word_embeddings}
\title{Double Anchor Word Embeddings}
\usage{
double_anchor_word_embeddings(anchoring_word, bwe_object,
  similarity = c("cosine", "weighted cosine"), trim = TRUE)
}
\arguments{
\item{anchoring_word}{Character. A single word, used to set the anchorings.}

\item{bwe_object}{List. Output from \code{fit_bwe}.}

\item{similarity}{Character. Should similarity be measured only by cosine similiarity, or should similarity be weighted by the log frequency of words.}

\item{trim}{Boolean. Should common stopwords be removed from the list of possible anchors? Setting this to \code{TRUE} improves performance and interpretability. Defaults to \code{TRUE}.}
}
\value{
A list. Contains a matrix of the anchored embeddings, and vector of the words used as anchors.
}
\description{
Anchors and identifies the embeddings produced from \code{bwe}. Unlike \code{anchor_word_embeddings}, endpoint are selected for each dimension.
}
