% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bwe.R
\name{fit_bwe}
\alias{fit_bwe}
\title{Estimate BWE}
\usage{
fit_bwe(dat, text_field = NA, context_window = 3L,
  minimum_word_count = 5L, K = 5, character_min = 0L,
  remove_numbers = FALSE, stops = NULL, start_type = c("svd",
  "zeros", "random"), verbose = TRUE, maxit = 5000, thresh = 0.01,
  checkfreq = 5, prop_ns = 5, cx_0 = 0.01, dx_0 = 0.01,
  cb_0 = 0.01, db_0 = 0.01, seed = NA)
}
\arguments{
\item{dat}{A \code{data.table} or \code{data.frame}, where every row is a document.}

\item{text_field}{Character string. The name of the column in \code{dat} which contains the text.}

\item{context_window}{Integer. The size of the word co-occurrence window. Defaults to 3.}

\item{minimum_word_count}{Integer. The minimum word frequency to be included in the corpus. Defaults to 5, but with larger corpora, increase this.}

\item{K}{Integer. The number of embedding dimensions. Defaults to five, but should usually be larger.}

\item{character_min}{Integer. What is the smallest sequence of characters to be considered? Anything smaller than this is discarded. Defaults to 0.}

\item{remove_numbers}{Logical. Argument to \code{quanteda} to remove numbers from the co-occurrence matrix. Defaults to \code{FALSE}.}

\item{stops}{Vector. A vector of stopwords to remove from the co-occurrence matrix. Defaults to \code{NULL}, with no words removed.}

\item{start_type}{Character. How the \code{x} and \code{beta} matrices are initialized. \code{zeroes} initialized with a matrix of zeroes, and \code{random} intializes with draws from a standard normal distribution. \code{svd} calculates the singular value decomposition of the co-occurrence matrix using \code{\link[irlba]{irlba}}, and initializes \code{x} and \code{beta} with the \code{u} and \code{v} values output from \code{irlba}. We recommend using this initializing, as it provides more useful initializations for the model.}

\item{verbose}{Boolean. Should model convergence be displayed as the model runs? Defaults to \code{TRUE}.}

\item{maxit}{Integer. Maximum number of iterations to run the model. Default is 5000.}

\item{thresh}{Numeric. Convergence threshold (Change in ELBO). Defaults to .01.}

\item{checkfreq}{Integer. How often should verbose output be displayed?}

\item{prop_ns}{Integer. Number of negative samples for every positive sample. Default is 5, but this should be adjusted depending on your dataset.}

\item{threads}{Integer. Number of cores to use, default is 1, threading done via OpenMP.}

\item{cx}{Positive numeric. Hyperparameters for \code{x} covariance.}

\item{dx}{Positive numeric. Hyperparameters for \code{x} covariance.}

\item{cb}{Positive numeric. Hyperparameters for \code{beta} covariance.}

\item{db}{Positive numeric. Hyperparameters for \code{beta} covariance.}
}
\value{
A list of objects

\item{data_prep}{A list of the inputs into the \code{bwe} function. Includes sparse co-occurrence matrix, word frequencies, transformed probabilities for negative sampling, and initialization values.}
\item{x}{The matrix of word embeddings. Rows are dimensions, columns are words.}
\item{beta}{The matrix of context embeddings. Rows are dimensions, columns are words.}
\item{var_x}{The covariance matrix for \code{x}.}
\item{var_beta}{The covariance matrix for \code{beta}.}
\item{xsigma}{The prior, learned from the data, for x.}
\item{betasigma}{The prior, learned from the data, for beta.}
\item{cx, dx, cb, bd}{The priors on \code{xsigma} and \code{betasigma}, learned from the data.}
\item{convergence}{A two column matrix. Column 1 is the rank of \code{x}, column two is the change in the elbo.}
\item{elbo}{A vector. The raw elbo values.}
}
\description{
Takes a collection of documents, converts them to a format for statistical estimation, and then estimates a Bayesian word embedding model using variational inference.
}
