% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_words.R
\name{nearest_words}
\alias{nearest_words}
\title{Return most similar words}
\usage{
nearest_words(vec1, mat1, n_similar = 10)
}
\arguments{
\item{vec1}{Vector. The vector for a word of interest.}

\item{mat1}{Matrix. The \code{x} matrix from either \code{fit_bwe} or \code{anchor_word_embeddings}.}

\item{n_similar}{Integer. Number of similar words to return. Defaults to 10}
}
\value{
A vector of length \code{n_similar}.
}
\description{
Return the most similar words to a given word, as determined by cosine similarity.
}
