% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_word_embeddings.R
\name{scaling_word_embeddings}
\alias{scaling_word_embeddings}
\title{Scales Word Embeddings}
\usage{
scaling_word_embeddings(target_word, bwe_object, trim = TRUE)
}
\arguments{
\item{target_word}{Character vector. A single word, used to set the scale the embeddings. If two words are provided, it is assumed they are opposites, they anchor each end of the embedding scaling.}

\item{bwe_object}{List. Output from \code{fit_bwe}.}

\item{trim}{Logical. Should common stopwords be removed from the list of scaled words? Setting this to \code{TRUE} improves performance and interpretability. Defaults to \code{TRUE}.}
}
\value{
A \code{data.table}. Contains each word, and its position relative to the scaling word(s).
}
\description{
Scales the word embeddings produced from \code{bwe} according to either one reference word, or a pair of opposite words.
}
